% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAI_query.R
\name{anaconda_ai_navigator_query}
\alias{anaconda_ai_navigator_query}
\title{anaconda_ai_navigator_query
Query Anaconda AI Navigator}
\usage{
anaconda_ai_navigator_query(
  prompt,
  model = NULL,
  history = NULL,
  api_key = NULL,
  base_url = "http://localhost",
  port = 8080,
  temperature = 0.8,
  top_k = 40,
  top_p = 0.95,
  n_predict = 200,
  stop = list("</s>"),
  ...
)
}
\arguments{
\item{prompt}{character, the query for the chatbot}

\item{model}{character, the model to use (see the `Models` pane in AI Navigator for appropriate options for your computer)}

\item{history}{list, a list of previous interactions in the conversation}

\item{api_key}{character, an API key for Anaconda AI Navigator (default = NULL)}

\item{base_url}{character, the base URL for the model (default = 'http://localhost')}

\item{port}{integer, the port on which the model is running (default = 8080)}

\item{temperature}{numeric, a value between 0 and 1 that controls the randomness of the model (default = 0.8)}

\item{top_k}{integer, the number of tokens to consider at each step (default = 40)}

\item{top_p}{numeric, the cumulative probability of tokens to consider at each step (default = 0.95)}

\item{n_predict}{integer, the number of tokens to predict (default = 200)}

\item{stop}{character, a list of tokens at which to stop (default = "</s>")}

\item{...}{Other parameters controlling model behavior (see details)}
}
\value{
character, The response from the chatbot.
}
\description{
anaconda_ai_navigator_query
Query Anaconda AI Navigator
}
\details{
This function queries a specific Anaconda AI Navigator model with the provided query.
AI Navigator does not require an API key to run locally.
When `api_key` is NULL, the function will assume the model is running locally and does not require an API key.

The output from these models can vary greatly depending on the model, prompt, and parameters.
Not all models utilize the `stop` parameter correctly, resulting in some rambling responses.
}
\seealso{
https://docs.anaconda.com/ai-navigator/
}
