% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_answer.R
\name{construct_preamble}
\alias{construct_preamble}
\title{construct_preamble
Construct the preamble for the query}
\usage{
construct_preamble(course = NULL, altPreamble = NULL)
}
\arguments{
\item{course}{character, The course title}

\item{altPreamble}{character, An alternative preamble to the default}
}
\value{
character, The preamble for the query
}
\description{
construct_preamble
Construct the preamble for the query
}
\examples{
construct_preamble()
construct_preamble('Data Science')
construct_preamble('Data Science', 'Is this correct? If not, just give me the answer.')
}
