% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{encrypt_text}
\alias{encrypt_text}
\alias{encrypt_text_one}
\alias{decrypt_text}
\alias{decrypt_text_one}
\title{encrypt_text
Encryption and decryption of text using the sodium library}
\usage{
encrypt_text(
  plain_text,
  key_hex = get_secret("ENCRYPTION_KEY", "Private", "learnr", "key_hex")
)

encrypt_text_one(
  plain_text,
  key_hex = get_secret("ENCRYPTION_KEY", "Private", "learnr", "key_hex")
)

decrypt_text(
  hex_cipher,
  key_hex = get_secret("ENCRYPTION_KEY", "Private", "learnr", "key_hex")
)

decrypt_text_one(
  hex_cipher,
  key_hex = get_secret("ENCRYPTION_KEY", "Private", "learnr", "key_hex")
)
}
\arguments{
\item{plain_text}{Character vector to encrypt}

\item{key_hex}{Encryption key}

\item{hex_cipher}{Character string to decrypt}
}
\value{
Encrypted text

Decrypted text
}
\description{
encrypt_text
Encryption and decryption of text using the sodium library

encrypt_text_one

decrypt_text
Decrypt text using the sodium library

decrypt_text_one
}
\details{
You will need to generate a 32-byte key (or load one, e.g. from `.Renviron`). See examples below.
}
\examples{
# set ENCRYPTION_KEY
key <- sodium::random(32)
Sys.setenv(ENCRYPTION_KEY = sodium::bin2hex(key))

# encrypt
(encrypted <- encrypt_text(c("Sensitive text to encrypt", "abc123")))

# decrypt
decrypt_text(encrypted)

}
