% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIkeys.R
\name{getAPIkey}
\alias{getAPIkey}
\title{getAPIkey
Retrieve an API key either using `op` or from the environment}
\usage{
getAPIkey(key_name, ...)
}
\arguments{
\item{key_name}{character name of the key to retrieve}

\item{...}{additional arguments passed to `op`}
}
\value{
character, the API key
}
\description{
getAPIkey
Retrieve an API key either using `op` or from the environment
}
\details{
This function retrieves an API key with the name specified by `key_name`.
If `key_name` is not an environmental variable, the function will attempt to retrieve the key from from 1Password using `op`.

To temporarily set the API key as an environmental variable (i.e. for the duration of R the session) use `Sys.setenv(key_name = "key")`, where "key" is the API key, and `key_name` is the name of the api key (e.g. "GEMINI_API_KEY").
To permanently set the API key as an environmental variable, add `key_name=key` (no quotes) to your `.Renviron` file - it is recommended you use `usethis::edit_r_envrion()` to open the .Renviron file if you are setting this globally.
}
\seealso{
`op`
}
