% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIkeys.R
\name{op}
\alias{op}
\title{op
Retrieve an item from 1Password using the `op` command line tool}
\usage{
op(...)
}
\arguments{
\item{...}{character, path to the item in 1Password}
}
\value{
character, the item from 1Password
}
\description{
op
Retrieve an item from 1Password using the `op` command line tool
}
\details{
This function retrieves an item from 1Password using the `op` command line tool.
The path to the item in 1Password should be provided as a series of arguments.
For example, to retrieve the API key, "api_key" from the "Gemini" item located in the "Private" vault, use `op("Private", "Gemini", "api_key")`.

This function works well on Unix-like systems (e.g. Linux, macOS), but may require additional configuration on Windows.
If this function is having trouble finding `op`, first verify that the cli command for `op` is installed (see https://developer.1password.com/docs/cli/get-started/).
If it is still having trouble, the `OP` environment variable can be set to the path to the `op` executable.
To find the install path on Windows, run `(Get-Command op).Source` in the power shell and copy the path returned (ending with "\" and excluding "op.exe").

To permanently add `OP` to the environment variables, open .Renviron with `usethis::edit_r_environ()` and add a new line with the something like the following (making sure you put "\\" for each backslash on Windows):
"OP=C:\\path\\to\\op\\" (Windows) or "OP=/path/to/op/" (Unix-like systems).
If `OP` isn't defined, `Sys.getenv("OP")` will return "", and the system call will look in the default PATH for `op`.
}
