% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gemini_response}
\alias{print.gemini_response}
\title{print.gemini_response
Print method for the response from `gemini_query`}
\usage{
\method{print}{gemini_response}(x, ...)
}
\arguments{
\item{x}{object, The response from the chatbot}

\item{...}{additional arguments}
}
\description{
print.gemini_response
Print method for the response from `gemini_query`
}
