% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_answer.R
\name{record_answer}
\alias{record_answer}
\title{record_answer
Record answers when submitted in a learnr tutorial}
\usage{
record_answer(
  record,
  db_path = Sys.getenv("SQLITE_DB_PATH", "submissions.sqlite"),
  table_name = "tutorial_submissions"
)
}
\arguments{
\item{record}{A data.frame containing `student_id`, `question_id`, `submission`, `is_correct`, `ai_feedback` and `timestamp`}

\item{db_path}{Path to the SQLite database}

\item{table_name}{Name of the table to add the data to}
}
\value{
Invisibly returns `TRUE` upon successful completion
}
\description{
record_answer
Record answers when submitted in a learnr tutorial
}
\details{
Assumes there is a correctly formatted database at `db_host`
}
