% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_answer.R
\name{check_answer}
\alias{check_answer}
\title{check_answer
Check the answer of the user}
\usage{
check_answer(
  question,
  answer,
  solution = NULL,
  preamble = construct_preamble()
)
}
\arguments{
\item{question}{character, The question}

\item{answer}{character, The answer of the user}

\item{solution}{character, The solution of the question}

\item{preamble}{character, The preamble for the query}
}
\value{
character, The response from the chatbot
}
\description{
check_answer
Check the answer of the user
}
