% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAI_query.R
\name{genAI_query}
\alias{genAI_query}
\title{genAI_query
Query a genAI chatbot}
\usage{
genAI_query(query, model = "gemini-1.5-flash", history = NULL, ...)
}
\arguments{
\item{query}{character, The query for the chatbot}

\item{model}{character, The model to use (see details)}

\item{history}{list, A list of previous interactions in the conversation}

\item{...}{Additional arguments passed to the function handling the query (i.e. `google_genAI`)}
}
\description{
genAI_query
Query a genAI chatbot
}
\details{
This function queries a specific model with the provided query. The package only supports Gemini at this time, but support for OpenAI will soon be added.
}
